showroom/usecase-004/requirements.txt
openai>=1.41.0
python-dotenv>=1.0.0
tabulate>=0.9.0